requirements.txt
pandas
requests
python-dotenv
azure-storage-blob